ValidateDataType function added
